Support for image and attachment parsing/posting & link parsing in debug window
Images and attachments can now be posted into Discord if they exist in the JSON file.
Links to images and attachments will now show in the debug window and are clickable. Slackord will launch the user's default browser to view images and attachments.
General code cleanup and minor performance improvements.