Update Slackord for Windows to v2.3
NEW - Slackord2 for Windows now allows selecting multiple JSON files via folder selection. The order of the files are based on hierarchical Windows default. It is best to rename multiple files in a folder based on the order you wish them to be posted.

This feature is not yet available for Linux and Mac, but I am working on it! I expect some bugs to come through, so before releasing them to both, it's best to squash the bugs found in the Windows version then move that code to Linux and Mac, as the codebase is quite a bit different now.

CHANGE - Upgraded Discord.NET library to the latest stable version (3.8.1).
CHANGE - Upgraded Octokit to library to the latest stable version (4.0.1)
CHANGE - Improved performance of Slackord2 greatly by reducing additional JSON parsing where possible.