v4.1.0
NEW - Slackord now uses webhooks and changes its name and image to the user that sent the message.

NEW - File attachments up to 25MB (the maximum size a non-Nitro user can send) are now supported. Any files over the limit will use the link to the attachment instead.

CHANGE - Since Slackord now uses webhooks to set the image and name of the bot to the post's user, messages now only contain the timestamp and the message contents. Previously, this was timstamp, user, messagecontent.