v4.2.0b1
NEW - File attachments are downloaded as part of reconstruction in the Downloads folder. If a file is larger than 25MB, the permalink to the file is instead used.

NEW - Slackord Logger. The logger now logs specific cases where failures occur and gives the full JSON for the object that caused an issue. This is to be used for support for Slackord going forward to help identify where issues are happening without users always needing to share their full JSON exports. Instead, the logfile at /Logs/Slackord_[DateTime].log can be shared. This logger will be expanded upon as new issues arise.

FIX - Timestamps are now converted to 12 or 24 hours based on the user's current culture format.

FIX - Multiple issues where a file attachment was being parsed for markdown, causing spam in the debug window.