v4.2.0b3
FIX - File downloads not working.
FIX - Maximum active threads being hit for larger imports.
FIX - Multiple cases where a file attachment would throw an exception:
```A message with a valid file that is not hidden by Slack.
A message with a file that is hidden by Slack and therefore not downloadable.
An empty message with a valid file.
An empty message with a file that is hidden by Slack.```